Update branch for codesandbox embed support (#1)
5 of the 6 code examples in this course work really well in codesandbox as it runs tests now.

The [Globals lesson](https://egghead.io/lessons/javascript-provide-helper-functions-as-globals) didn't work with codesandbox well. I did get it to work but the preview panel was erroring out with a 502 bad gateway. Though the tests ran as expected in the terminal. I plan on just linking the GitHub repository for this lesson unless you have an idea of how to better run that example in codesandbox

## Changes Made
- Updated lesson files to ordered lesson folder
- Added a `package.json` and `index.html` file for each example that will work in codesandbox

![](https://media.giphy.com/media/gw3IWyGkC0rsazTi/giphy.gif)